smartflow_backend/core/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
alembic==1.12.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
requests==2.31.0
httpx==0.24.1
websockets==11.0.3
redis==5.0.0
celery==5.3.4
pydantic[email]==2.3.0 